COVID/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bkcharts==0.2
bleach==3.0.2
bokeh==2.0.0
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
conda==4.5.11
cryptography==2.3.1
cycler==0.10.0
Cython==0.29
dask==2.12.0
decorator==4.3.0
defusedxml==0.5.0
dill==0.3.1.1
distributed==2.12.0
dominate==2.5.1
entrypoints==0.2.3
feather-format==0.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
future==0.18.2
gast==0.2.2
gitdb==4.0.2
GitPython==3.1.0
google-pasta==0.1.5
grpcio==1.19.0
h5py==2.9.0
HeapDict==1.0.1
idna==2.7
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.11.1
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.19.6
kiwisolver==1.0.1
lesscpy==0.13.0
Markdown==3.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.4
msgpack==1.0.0
multiprocess==0.70.9
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
numpy==1.18.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.3.1
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
ply==3.11
prometheus-client==0.4.2
prompt-toolkit==2.0.6
protobuf==3.7.1
psutil==5.7.0
ptyprocess==0.6.0
pyabc==0.10.1
pyarrow==0.16.0
pycosat==0.6.3
pycparser==2.18
Pygments==2.2.0
pyOpenSSL==18.0.0
pyparsing==2.2.2
PySocks==1.6.8
python-dateutil==2.7.3
pytz==2018.5
PyYAML==5.3.1
pyzmq==17.1.2
qtconsole==4.4.2
redis==3.4.1
requests==2.19.1
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smmap==3.0.1
sortedcontainers==2.1.0
SQLAlchemy==1.3.15
tblib==1.6.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.2
typing-extensions==3.7.4.1
urllib3==1.23
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
zict==2.0.0
